Bug: 29977 Ajuste query ObterAvaliacoesBimestrais.

+++ src/SME.SGP.Dados/Repositorios/RepositorioAtividadeAvaliativaDisciplina.cs
@@ -46,7 +46,7 @@ namespace SME.SGP.Dados.Repositorios
                        inner join tipo_avaliacao t on t.id = aa.tipo_avaliacao_id and t.codigo = 1
                        inner join periodo_escolar p on p.periodo_inicio <= aa.data_avaliacao and p.periodo_fim >= aa.data_avaliacao 
                        where not aa.excluido 
-                        and aad.disciplina_id = @disciplinaId
+                        and aad.disciplina_id = @disciplinaId::int8
                         and aa.turma_id = @turmaId
                         and p.bimestre = @bimestre
                         and p.tipo_calendario_id = @tipoCalendarioId ";
